data_ingestion/requirements.txt
pymongo
python-dotenv
tqdm
sentence-transformers